# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing LiteLLM keys and user budgets.",
        "title": "LLM Request Manager API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/keys": {
            "post": {
                "description": "Create a new API key with optional budget and type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keys"
                ],
                "summary": "Create a new API key",
                "parameters": [
                    {
                        "description": "Create Key Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.GenerateKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/keys/active": {
            "get": {
                "description": "Fetch active keys from LiteLLM and sync with local DB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keys"
                ],
                "summary": "Get active keys",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.ActiveKeyResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/keys/history": {
            "get": {
                "description": "Fetch revoked or deleted keys from local DB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keys"
                ],
                "summary": "Get key history",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.KeyHistory"
                            }
                        }
                    }
                }
            }
        },
        "/keys/{key_id}": {
            "delete": {
                "description": "Revoke an API key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keys"
                ],
                "summary": "Delete an API key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Key ID",
                        "name": "key_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/me": {
            "get": {
                "description": "Fetch current user information from LiteLLM",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get current user info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.LiteLLMUser"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ActiveKeyResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "key_id": {
                    "type": "string"
                },
                "mask": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "spend": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "handlers.CreateKeyRequest": {
            "type": "object",
            "properties": {
                "budget": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "description": "\"standard\" or \"long-term\"",
                    "type": "string"
                }
            }
        },
        "models.KeyHistory": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "keyMask": {
                    "type": "string"
                },
                "keyName": {
                    "type": "string"
                },
                "keyType": {
                    "type": "string"
                },
                "liteLLMKeyID": {
                    "type": "string"
                },
                "revokedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "services.GenerateKeyResponse": {
            "type": "object",
            "properties": {
                "hidden": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "key_alias": {
                    "type": "string"
                },
                "key_name": {
                    "type": "string"
                },
                "max_budget": {
                    "type": "number"
                },
                "spend": {
                    "type": "number"
                },
                "token": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "services.LiteLLMUser": {
            "type": "object",
            "properties": {
                "max_budget": {
                    "type": "number"
                },
                "spend": {
                    "type": "number"
                },
                "user_email": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}